+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Set defaults for file and line args of Violation#new & Markdown#new - Juanito Fatas
 * Add docs to `String#danger_pluralize` & `String#danger_underscore` - Juanito Fatas
 * Refactor Comment class out of CommentsHelper - Juanito Fatas
+* Add support for Bitbucket Cloud - Wolfgang Damm
 
 ## 3.2.2